+++ handlers.js
@@ -9,27 +9,31 @@ const main = require('./index');
 const swCommands = require('./modules/SW.GENESYS/').commands;
 const l5rCommands = require('./modules/L5R/').commands;
 
-// Called every time the bot connects to Twitch chat:
-const onReady = client => {
-    console.log(`Logged in as ${client.user.username}!`);
-    console.log(`Version: ${version}`);
-};
-
 //Called whenever a users send a message to the server
-const onMessage = async (message, client) => {
-    let prefix, params, command, desc, channelEmoji;
-
+const onMessage = async ({ message, client }) => {
     //Ignore messages sent by the bot
     if (message.author.bot) return;
 
     //check to see if bot can send messages on channel and external emoji can be used
-    if (message.channel.type !== 'dm') {
+    if (message.channel.type !== 'DM') {
         if (!message.channel.permissionsFor(client.user).has('SEND_MESSAGES')) return;
-        // if (!message.channel.permissionsFor(client.user).has('USE_EXTERNAL_EMOJIS')) {
-        //     main.sendMessage(message, `Please enable \'Use External Emoji\' for ${client.user.username}`);
-        //     return;
-        // }
+        if (!message.channel.permissionsFor(client.user).has('USE_EXTERNAL_EMOJIS')) {
+            await main.sendMessage({
+                message,
+                text: `Please enable \'Use External Emoji\' permission for ${client.user.username}`
+            });
+            return;
+        }
+        if (!message.channel.permissionsFor(client.user).has('EMBED_LINKS')) {
+            await main.sendMessage({
+                message,
+                text: `Please enable \'Embed Links\' permission for ${client.user.username}`
+            });
+            return;
+        }
     }
+
+    let prefix, params, command, desc, channelEmoji;
 
     //build the prefix
     prefix = await buildPrefix(client, message);
@@ -44,18 +48,13 @@ const onMessage = async (message, client) => {
     if (!command || command === 'play') return;
 
     //get channelEmoji
-    channelEmoji = main.channelEmoji[message.channel.id];
     if (!channelEmoji) channelEmoji = await readData(client, message, 'channelEmoji').catch(console.error);
-    main.addChannelEmoji(message, channelEmoji);
 
     //check for Patron
-    if (config.patronDiceRole && config.patreonGuild && config[`${channelEmoji}Patreon`]) {
-        if (await modules.checkPatreon(client, message.author.id)) channelEmoji += 'Patreon';
-        if (!channelEmoji.includes('Patreon') && config.patronMegaRole && message.guild) {
-            if (await modules.checkPatreonServer(client, message.guild.ownerID)) {
-                channelEmoji += 'Patreon';
-            }
-        }
+    const member = await main.checkWithBabyBot(message.author.id);
+
+    if (member) {
+        channelEmoji += 'Patreon';
     }
 
     //make the descriptor
@@ -69,13 +68,13 @@ const onMessage = async (message, client) => {
 
 //************************COMMANDS START HERE************************
 
-    switch(command) {
+    switch (command) {
         case 'stats':
-            modules.buildStats(client, message);
+            await modules.stats({ client, message });
             break;
         case 'ver':
         case 'v':
-            main.sendMessage(message, `${client.user.username}: version: ${version}`);
+            await main.sendMessage({ message, text: `${client.user.username}: version: ${version}` });
             break;
         case 'poly':
         case 'p':
@@ -85,35 +84,36 @@ const onMessage = async (message, client) => {
         case 'genesys':
         case 'l5r':
             writeData(client, message, 'channelEmoji', command);
-            main.addChannelEmoji(message, command);
-            main.sendMessage(message, `${client.user.username} will now use ${command} dice`);
+            await main.sendMessage({
+                message, text: `${client.user.username} will now use ${command} dice`
+            });
             break;
         case 'prefix':
             if (message.channel.type === 'dm') {
-                main.sendMessage(message, 'Prefix cannot be changed in DMs');
+                await main.sendMessage({ message, text: 'Prefix cannot be changed in DMs' });
             } else await modules.prefix(client, message, params);
             break;
         case 'invite':
             const embed = new Discord.MessageEmbed()
-                .setColor('777777')
+                .setColor('GREY')
                 .setTitle(`**Invite**`)
-                .setDescription(`Click [here](https://discordapp.com/oauth2/authorize?client_id=${client.user.id}&scope=bot&permissions=280576) to invite the bot to your server!`);
-            main.sendMessage(message, { embed });
+                .setDescription(`Click [here](https://discordapp.com/oauth2/authorize?client_id=${client.user.id}&scope=bot&permissions=105227020288&scope=bot%20applications.commands) to invite the bot to your server!`);
+            await main.sendMessage({ message, embed });
             break;
     }
     if (message.author.id === config.adminID) {
-        await modules.admin(client, message, params, command);
+        await modules.admin({ client, message, params, command });
     }
-    switch(channelEmoji) {
+    switch (channelEmoji) {
         case 'swrpg':
         case 'swrpgPatreon':
         case 'genesys':
         case 'genesysPatreon':
-            swCommands(client, message, params, command, desc, channelEmoji, prefix);
+            await swCommands({ client, message, params, command, desc, channelEmoji, prefix });
             break;
         case 'l5r':
         case 'l5rPatreon':
-            l5rCommands(client, message, params, command, desc, channelEmoji, prefix);
+            await l5rCommands({ client, message, params, command, desc, channelEmoji, prefix });
             break;
         default:
             break;
@@ -121,5 +121,4 @@ const onMessage = async (message, client) => {
     }
 };
 
-exports.onReady = onReady;
 exports.onMessage = onMessage;