+++ modules/L5R/help.js
@@ -2,7 +2,7 @@ const {MessageEmbed} = require('discord.js');
 const main = require('../../index');
 
 const help = (topic, message, prefix) => {
-	const embed = new MessageEmbed().setColor('ab0f1a');
+	const embed = new MessageEmbed().setColor('DARK_NAVY');
 	switch (topic) {
 		case `roll`:
 		case 'r':
@@ -46,8 +46,8 @@ const help = (topic, message, prefix) => {
 				.addField('Role playing games by Fantasy Flight Games', `[Edge of the Empire](https://www.fantasyflightgames.com/en/products/star-wars-edge-of-the-empire), [Force and Destiny](https://www.fantasyflightgames.com/en/products/star-wars-force-and-destiny), [Age of Rebellion](https://www.fantasyflightgames.com/en/products/star-wars-age-ofrebellion),[Genesys](https://www.fantasyflightgames.com/en/products/genesys), [Legends of the Five Rings](https://www.fantasyflightgames.com/en/legend-of-the-five-rings-roleplaying-game)`);
 			break;
 	}
-	main.sendMessage(message, {embed});
+	main.sendMessage({message, embed});
 
 }
 
-exports.help = help;
+module.exports = help;