+++ modules/L5R/index.js
@@ -1,4 +1,4 @@
-exports.commands = require('./commands').commands;
-exports.help = require('./help').help;
+exports.commands = require('./commands');
+exports.help = require('./help');
 exports.roll = require('./roll').roll;
 exports.keep = require('./roll').keep;