+++ modules/SW.GENESYS/initiative.js
@@ -2,7 +2,7 @@ let functions = require('./');
 const { readData, writeData } = require('../data');
 const main = require('../../index');
 
-const initiative = async (client, message, params, channelEmoji) => {
+const initiative = async ({ client, message, params, channelEmoji }) => {
     let initiativeOrder = await readData(client, message, 'initiativeOrder');
     let merge = false;
     let write = true;
@@ -10,21 +10,21 @@ const initiative = async (client, message, params, channelEmoji) => {
     if (!initiativeOrder.newslots) initiativeOrder.newslots = [];
     if (!initiativeOrder.slots) initiativeOrder.slots = [];
     let command = params.shift();
-    switch(command) {
+    switch (command) {
         //roll for initiativeOrder
         case 'roll':
         case 'r':
             merge = true;
             if (!params[0] || params[0] === 'npc' || params[0] === 'pc') {
-                main.sendMessage(message, 'No dice defined.  ie \'!init roll yygg npc/pc\'');
+                main.sendMessage({ message, text: 'No dice defined.  ie \'!init roll yygg npc/pc\'' });
                 return;
             }
             if (!(params[params.length - 1] === 'npc' || params[params.length - 1] === 'pc')) {
-                main.sendMessage(message, 'No Character type defined.  ie \'!init roll yygg npc/pc\'');
+                main.sendMessage({ message, text: 'No Character type defined.  ie \'!init roll yygg npc/pc\'' });
                 return;
             }
             let type = params.pop();
-            let diceResult = await functions.roll(client, message, params, channelEmoji, 'Initiative roll');
+            let diceResult = await functions.roll({ client, message, params, channelEmoji, desc: 'Initiative roll' });
             diceResult = diceResult.results;
             let rollResult = {
                 success: diceResult.success,
@@ -35,10 +35,16 @@ const initiative = async (client, message, params, channelEmoji) => {
             if (initiativeOrder.turn !== 1) {
                 initiativeOrder.newslots.push(rollResult);
                 if (type === 'npc') {
-                    main.sendMessage(message, ':smiling_imp: will be added to the initiative order in the next round')
+                    main.sendMessage({
+                        message,
+                        text: ':smiling_imp: will be added to the initiative order in the next round'
+                    });
                 }
                 if (type === 'pc') {
-                    main.sendMessage(message, ':slight_smile: will be added to the initiative order in the next round')
+                    main.sendMessage({
+                        message,
+                        text: ':slight_smile: will be added to the initiative order in the next round'
+                    });
                 }
             } else {
                 initiativeOrder.slots.push(rollResult);
@@ -49,11 +55,11 @@ const initiative = async (client, message, params, channelEmoji) => {
         case 's':
             initiativeOrder = initializeInitOrder();
             if (!params[0]) {
-                main.sendMessage(message, 'No Initiative Order defined.  ie \'!init set nppnn\'');
+                main.sendMessage({ message, text: 'No Initiative Order defined.  ie \'!init set nppnn\'' });
                 break;
             }
-            for(let i = 0; i < params[0].length; i++) {
-                switch(params[0][i]) {
+            for (let i = 0; i < params[0].length; i++) {
+                switch (params[0][i]) {
                     case 'n':
                         initiativeOrder.slots.push({ type: 'npc' });
                         break;
@@ -76,7 +82,7 @@ const initiative = async (client, message, params, channelEmoji) => {
             if (initiativeOrder.turn + 1 > initiativeOrder.slots.length) {
                 initiativeOrder.turn = 1;
                 initiativeOrder.round++;
-                main.sendMessage(message, 'New Round!');
+                main.sendMessage({ message, text: 'New Round!' });
                 if (initiativeOrder.newslots.length > 0) {
                     initiativeOrder.slots = initiativeOrder.slots.concat(initiativeOrder.newslots);
                     initiativeOrder.newslots = [];
@@ -87,23 +93,23 @@ const initiative = async (client, message, params, channelEmoji) => {
         case 'previous':
         case 'p':
             if (initiativeOrder.turn === 1 && initiativeOrder.round === 1) {
-                main.sendMessage(message, 'Initiative is already at the starting turn!');
+                main.sendMessage({ message, text: 'Initiative is already at the starting turn!' });
             } else if (initiativeOrder.turn - 1 < 1) {
                 initiativeOrder.turn = initiativeOrder.slots.length;
                 initiativeOrder.round--;
-                main.sendMessage(message, 'Previous Round!');
+                main.sendMessage({ message, text: 'Previous Round!' });
             } else initiativeOrder.turn--;
             break;
         //manually modify the initiativeOrder
         case 'modify':
             //check if numbers are used
             if (!params[0]) {
-                main.sendMessage(message, 'No Initiative Order defined.  ie \'!init set nppnn\'');
+                main.sendMessage({ message, text: 'No Initiative Order defined.  ie \'!init set nppnn\'' });
                 break;
             }
             initiativeOrder.slots = [];
-            for(let i = 0; i < params[0].length; i++) {
-                switch(params[0][i]) {
+            for (let i = 0; i < params[0].length; i++) {
+                switch (params[0][i]) {
                     case 'n':
                         initiativeOrder.slots.push({ type: 'npc' });
                         break;
@@ -129,11 +135,11 @@ const initiative = async (client, message, params, channelEmoji) => {
             write = false;
             break;
     }
-    if(write) {
+    if (write) {
         writeData(client, message, 'initiativeOrder', initiativeOrder, merge);
     }
     if (initiativeOrder.slots[0]) printInitiativeOrder(initiativeOrder, message);
-    else main.sendMessage(message, 'No initiative order is set!');
+    else main.sendMessage({ message, text: 'No initiative order is set!' });
 };
 
 //Adds a roll to the order and sorts it
@@ -165,29 +171,31 @@ const initializeInitOrder = () => {
         slots: [],
         newslots: []
     };
-}
+};
 
 //Prints out Initiative Order to channel
 const printInitiativeOrder = (initiativeOrder, message) => {
     if (Object.keys(initiativeOrder.slots[0]).length > 1) initiativeOrder = sortInitiativeOrder(initiativeOrder);
     let faces = '';
-    for(let i = initiativeOrder.turn - 1; i < initiativeOrder.slots.length; i++) {
+    for (let i = initiativeOrder.turn - 1; i < initiativeOrder.slots.length; i++) {
         faces += getFace(initiativeOrder.slots[i].type);
     }
     faces += ':repeat:';
-    for(let i = 0; i < initiativeOrder.turn - 1; i++) {
+    for (let i = 0; i < initiativeOrder.turn - 1; i++) {
         faces += getFace(initiativeOrder.slots[i].type);
     }
-    main.sendMessage(message, 'Round: ' + initiativeOrder.round + ' Turn: ' + initiativeOrder.turn + ' Initiative Order: \n')
-           ;
+    main.sendMessage({
+        message,
+        text: 'Round: ' + initiativeOrder.round + ' Turn: ' + initiativeOrder.turn + ' Initiative Order: \n'
+    });
     if (faces === '') return;
     if (faces.length > 1500) faces = `Initiative order too long to display.`;
 
-    main.sendMessage(message, faces);
-}
+    main.sendMessage({ message, text: faces });
+};
 
 const getFace = (type) => {
-    switch(type) {
+    switch (type) {
         case 'npc': // non-playable character
             return ':smiling_imp:';
         case 'pc': // playable character
@@ -195,6 +203,6 @@ const getFace = (type) => {
         default:
             return ''; // Always return a string. Even an empty one.
     }
-}
+};
 
-exports.initiative = initiative;
+module.exports = initiative;