+++ modules/SW.GENESYS/trigger.js
@@ -1,11 +1,11 @@
 const { dice, readData } = require('../');
 const main = require('../../index');
 
-const trigger = async (client, message, type) => {
+const trigger = async ({ client, message, type }) => {
     let characterStatus = await readData(client, message, 'characterStatus');
     let list = [];
     if (Object.keys(characterStatus).length === 0) {
-        main.sendMessage(message, 'No characters found please use !char to setup');
+        main.sendMessage({ message, text: 'No characters found please use !char to setup' });
         return;
     }
     Object.keys(characterStatus).forEach(characterName => {
@@ -24,20 +24,20 @@ const trigger = async (client, message, type) => {
     let target = 0;
     let total = 0;
     list.forEach(name => total += name.value);
-    main.sendMessage(message, `The total group ${type} is ${total}. The ${type} roll is ${roll}.`);
+    main.sendMessage({ message, text: `The total group ${type} is ${total}. The ${type} roll is ${roll}.` });
 
     if (roll > total) {
-        main.sendMessage(message, `No ${type} triggered`);
+        main.sendMessage({ message, text: `No ${type} triggered` });
         return;
     }
 
     for (let i = 0; i < list.length; i++) {
         target += list[i].value;
         if (target > roll) {
-            main.sendMessage(message, `${list[i].name}'s ${list[i][type]} ${type} has been triggered.`);
+            main.sendMessage({ message, text: `${list[i].name}'s ${list[i][type]} ${type} has been triggered.` });
             break;
         }
     }
 };
 
-exports.trigger = trigger;
+module.exports = trigger;