+++ modules/emoji.js
@@ -1,0 +1,5 @@
+const emojiFile = require('../emoji.json');
+
+const emoji = (string, type = 'swrpg') => emojiFile[type][string];
+
+module.exports = emoji;